Model Permissions

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -3,12 +3,18 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use DB;
+use App\User;
 use App\Project;
+use App\ItemPermission;
 use Illuminate\Http\Request;
+use Spatie\Permission\Models\Permission;
 
+use App\Services\Permissions\ModelPermission;
 
 class ProjectController extends Controller
 {
+    private $check;
     /**
      * Display a listing of the resource.
      *
@@ -16,12 +22,12 @@ class ProjectController extends Controller
      */
     function __construct()
     {
-        $this->middleware('permission:project-list');
-        $this->middleware('permission:project-create', ['only' => ['create','store']]);
-        $this->middleware('permission:project-edit', ['only' => ['edit','update']]);
+        // $this->middleware('permission:project-list');
+        // $this->middleware('permission:project-create', ['only' => ['create','store']]);
+        // $this->middleware('permission:project-edit', ['only' => ['edit','update']]);
+        // $this->middleware('permission:project-delete', ['only' => ['destroy']]);
 
-
-        $this->middleware('permission:project-delete', ['only' => ['destroy']]);
+        $this->check = new ModelPermission();
     }
     /**
      * Display a listing of the resource.
@@ -30,6 +36,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
+
         $projects = Project::latest()->paginate(5);
         return view('admin.projects.index',compact('projects'))
             ->with('i', (request()->input('page', 1) - 1) * 5);
@@ -60,8 +67,32 @@ class ProjectController extends Controller
             'detail' => 'required',
         ]);
 
+        $project = Project::create($request->all());
 
-        Project::create($request->all());
+        $permissionNameShow = 'project-show-'.$project->id;
+        $permissionNameEdit = 'project-edit-'.$project->id;
+        $permissionNameDelete = 'project-delete-'.$project->id;
+
+        ItemPermission::create([
+            'type' => 'show',
+            'name' => $permissionNameShow,
+            'model_name' => 'App/Project',
+            'model_id' => $project->id
+        ]);
+
+        ItemPermission::create([
+            'type' => 'edit',
+            'name' => $permissionNameEdit,
+            'model_name' => 'App/Project',
+            'model_id' => $project->id
+        ]);
+
+        ItemPermission::create([
+            'type' => 'delete',
+            'name' => $permissionNameDelete,
+            'model_name' => 'App/Project',
+            'model_id' => $project->id
+        ]);
 
 
         return redirect()->route('admin.projects.index')
@@ -72,11 +103,17 @@ class ProjectController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Project  $project
      * @return \Illuminate\Http\Response
      */
     public function show(Project $project)
     {
+        $showOptions = array('type' => 'show', 'model_name' => 'App/Project', 'model_id' => $project->id);
+        $this->check->handle('project-list', $showOptions);
+        $usersWithRoles = $this->check->authUsers('App/Project', $project->id);
+
+        dump($usersWithRoles);
+
         return view('admin.projects.show',compact('project'));
     }
 
@@ -84,11 +121,14 @@ class ProjectController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Project  $project
      * @return \Illuminate\Http\Response
      */
     public function edit(Project $project)
     {
+        $editOptions = array('type' => 'edit', 'model_name' => 'App/Project', 'model_id' => $project->id);
+        $this->check->handle('project-edit', $editOptions);
+
         return view('admin.projects.edit',compact('project'));
     }
 
@@ -97,11 +137,14 @@ class ProjectController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Product  $product
+     * @param  \App\Project  $product
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Project $project)
     {
+        $editOptions = array('type' => 'edit', 'model_name' => 'App/Project', 'model_id' => $project->id);
+        $this->check->handle('project-edit', $editOptions);
+
         request()->validate([
             'name' => 'required',
             'detail' => 'required',
@@ -119,15 +162,18 @@ class ProjectController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Product  $product
+     * @param  \App\Project  $product
      * @return \Illuminate\Http\Response
      */
     public function destroy(Project $project)
     {
-        $project->delete();
+        $deleteOptions = array('type' => 'delete', 'model_name' => 'App/Project', 'model_id' => $project->id);
+        $this->check->handle('project-delete', $deleteOptions);
 
+        $project->delete();
+        ItemPermission::where(['model_id' => $project->id, 'model_name' => 'App/Project'])->delete();
 
         return redirect()->route('admin.projects.index')
-            ->with('success','Product deleted successfully');
+            ->with('success','Project deleted successfully');
     }
 }
